+++ Controllers/ConfigController.cs
@@ -4,25 +4,13 @@ using System.Text;
 using System.Threading.Tasks;
 using JacRed;
 using JacRed.Configuration.Schema;
+using JacRed.Models;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 namespace JacRed.Controllers
 {
-    /// <summary>Request body for config validate, diff, and save.</summary>
-    public class ConfigSaveRequest
-    {
-        /// <summary>Raw init.yaml / init.conf text.</summary>
-        public string content { get; set; }
-
-        /// <summary>yaml or json — format hint for content or output file.</summary>
-        public string format { get; set; }
-
-        /// <summary>Parsed configuration object (settings form UI).</summary>
-        public JObject data { get; set; }
-    }
-
     /// <summary>
     /// Configuration management API (init.yaml / init.conf).
     /// LAN or X-Dev-Key when devkey is configured.
