+++ Infrastructure/Trackers/Bitru/BitruSyncService.cs
@@ -22,8 +22,8 @@ namespace JacRed.Infrastructure.Trackers.Bitru
         static Dictionary<string, List<TaskParse>> taskParse = new Dictionary<string, List<TaskParse>>();
 
         static readonly TrackerParseLock _parseLock = new TrackerParseLock();
-        static volatile bool _parseAllTaskWork;
-        static readonly SemaphoreSlim _parseLatestSemaphore = new SemaphoreSlim(1, 1);
+        static readonly TrackerWorkFlag _parseAllTaskWork = new TrackerWorkFlag();
+        static readonly TrackerLatestParseLock _parseLatestLock = new TrackerLatestParseLock();
 
         static BitruSyncService()
         {
@@ -94,12 +94,7 @@ namespace JacRed.Infrastructure.Trackers.Bitru
 
         public async Task<string> ParseAllTaskAsync(CancellationToken cancellationToken = default)
         {
-            if (_parseAllTaskWork)
-                return "work";
-
-            _parseAllTaskWork = true;
-
-            try
+            return await TrackerSyncHelpers.RunParseAllTaskAsync(TrackerName, _parseAllTaskWork, checkDisabled: false, async () =>
             {
                 foreach (var task in taskParse.ToArray())
                 {
@@ -115,67 +110,58 @@ namespace JacRed.Infrastructure.Trackers.Bitru
                             val.updateTime = DateTime.Today;
                     }
                 }
-            }
-            catch { }
-
-            _parseAllTaskWork = false;
-            return "ok";
+            }, cancellationToken);
         }
 
         public async Task<string> ParseLatestAsync(int pages = 5, CancellationToken cancellationToken = default)
         {
-            if (!await _parseLatestSemaphore.WaitAsync(0, cancellationToken))
-                return "work";
-
-            var log = new StringBuilder();
-
-            try
+            return await TrackerSyncHelpers.RunParseLatestAsync(TrackerName, _parseLatestLock, checkDisabled: false, async () =>
             {
-                var sw = Stopwatch.StartNew();
-                ParserLog.Write(TrackerName, $"Starting ParseLatest pages={pages}");
+                var log = new StringBuilder();
 
-                foreach (var task in taskParse.ToArray())
+                try
                 {
-                    var pagesToParse = task.Value.OrderBy(x => x.page).Take(pages).ToArray();
+                    var sw = Stopwatch.StartNew();
+                    ParserLog.Write(TrackerName, $"Starting ParseLatest pages={pages}");
 
-                    foreach (var val in pagesToParse)
+                    foreach (var task in taskParse.ToArray())
                     {
-                        await Task.Delay(AppInit.conf.Bitru.parseDelay, cancellationToken);
+                        var pagesToParse = task.Value.OrderBy(x => x.page).Take(pages).ToArray();
 
-                        bool res = await BitruParser.ParsePageAsync(task.Key, val.page);
-                        if (res)
+                        foreach (var val in pagesToParse)
                         {
-                            val.updateTime = DateTime.Today;
-                            log.AppendLine($"{task.Key} - {val.page}");
+                            await Task.Delay(AppInit.conf.Bitru.parseDelay, cancellationToken);
+
+                            bool res = await BitruParser.ParsePageAsync(task.Key, val.page);
+                            if (res)
+                            {
+                                val.updateTime = DateTime.Today;
+                                log.AppendLine($"{task.Key} - {val.page}");
+                            }
                         }
                     }
+
+                    ParserLog.Write(TrackerName, $"ParseLatest completed successfully (took {sw.Elapsed.TotalSeconds:F1}s)");
+                }
+                catch (System.Net.Http.HttpRequestException ex)
+                {
+                    ParserLog.Write(TrackerName, $"ParseLatest HTTP Error: {ex.Message}");
+                }
+                catch (TaskCanceledException ex)
+                {
+                    ParserLog.Write(TrackerName, $"ParseLatest Cancelled: {ex.Message}");
+                }
+                catch (InvalidOperationException ex)
+                {
+                    ParserLog.Write(TrackerName, $"ParseLatest Invalid Operation: {ex.Message}");
+                }
+                catch (Exception ex)
+                {
+                    ParserLog.Write(TrackerName, $"ParseLatest Unexpected Error: {ex.Message}");
                 }
 
-                ParserLog.Write(TrackerName, $"ParseLatest completed successfully (took {sw.Elapsed.TotalSeconds:F1}s)");
-            }
-            catch (System.Net.Http.HttpRequestException ex)
-            {
-                ParserLog.Write(TrackerName, $"ParseLatest HTTP Error: {ex.Message}");
-            }
-            catch (TaskCanceledException ex)
-            {
-                ParserLog.Write(TrackerName, $"ParseLatest Cancelled: {ex.Message}");
-            }
-            catch (InvalidOperationException ex)
-            {
-                ParserLog.Write(TrackerName, $"ParseLatest Invalid Operation: {ex.Message}");
-            }
-            catch (Exception ex)
-            {
-                ParserLog.Write(TrackerName, $"ParseLatest Unexpected Error: {ex.Message}");
-            }
-            finally
-            {
-                _parseLatestSemaphore.Release();
-            }
-
-            var logText = log.ToString();
-            return string.IsNullOrWhiteSpace(logText) ? "ok" : logText;
+                return log.ToString();
+            }, cancellationToken);
         }
     }
 }
