+++ Infrastructure/Trackers/Kinozal/KinozalSyncService.cs
@@ -29,8 +29,8 @@ namespace JacRed.Infrastructure.Trackers.Kinozal
         static string Cookie;
 
         static readonly TrackerParseLock _parseLock = new TrackerParseLock();
-        static bool _parseAllTaskWork = false;
-        static readonly SemaphoreSlim _parseLatestSemaphore = new SemaphoreSlim(1, 1);
+        static readonly TrackerWorkFlag _parseAllTaskWork = new TrackerWorkFlag();
+        static readonly TrackerLatestParseLock _parseLatestLock = new TrackerLatestParseLock();
 
         static readonly List<string> Categories = new List<string>()
         {
@@ -196,12 +196,7 @@ namespace JacRed.Infrastructure.Trackers.Kinozal
 
         public async Task<string> ParseAllTaskAsync()
         {
-            if (_parseAllTaskWork)
-                return "work";
-
-            _parseAllTaskWork = true;
-
-            try
+            return await TrackerSyncHelpers.RunParseAllTaskAsync(TrackerName, _parseAllTaskWork, checkDisabled: false, async () =>
             {
                 foreach (var cat in taskParse.ToArray())
                 {
@@ -220,59 +215,49 @@ namespace JacRed.Infrastructure.Trackers.Kinozal
                         }
                     }
                 }
-            }
-            catch { }
-
-            _parseAllTaskWork = false;
-            return "ok";
+            });
         }
 
         public async Task<string> ParseLatestAsync(int pages = 5)
         {
-            if (!await _parseLatestSemaphore.WaitAsync(0))
-                return "work";
-
-            var log = new StringBuilder();
-
-            try
+            return await TrackerSyncHelpers.RunParseLatestAsync(TrackerName, _parseLatestLock, checkDisabled: false, async () =>
             {
-                var sw = Stopwatch.StartNew();
-                ParserLog.Write(TrackerName, $"Starting ParseLatest pages={pages}");
+                var log = new StringBuilder();
 
-                foreach (var cat in taskParse.ToArray())
+                try
                 {
-                    foreach (var arg in cat.Value.ToArray())
+                    var sw = Stopwatch.StartNew();
+                    ParserLog.Write(TrackerName, $"Starting ParseLatest pages={pages}");
+
+                    foreach (var cat in taskParse.ToArray())
                     {
-                        // Get first N pages sorted by page number
-                        var pagesToParse = arg.Value.OrderBy(x => x.page).Take(pages).ToArray();
-
-                        foreach (var val in pagesToParse)
+                        foreach (var arg in cat.Value.ToArray())
                         {
-                            await Task.Delay(AppInit.conf.Kinozal.parseDelay);
+                            var pagesToParse = arg.Value.OrderBy(x => x.page).Take(pages).ToArray();
 
-                            bool res = await parsePage(cat.Key, val.page, arg.Key);
-                            if (res)
+                            foreach (var val in pagesToParse)
                             {
-                                val.updateTime = DateTime.Today;
-                                log.AppendLine($"{cat.Key} - {arg.Key} - {val.page}");
+                                await Task.Delay(AppInit.conf.Kinozal.parseDelay);
+
+                                bool res = await parsePage(cat.Key, val.page, arg.Key);
+                                if (res)
+                                {
+                                    val.updateTime = DateTime.Today;
+                                    log.AppendLine($"{cat.Key} - {arg.Key} - {val.page}");
+                                }
                             }
                         }
                     }
+
+                    ParserLog.Write(TrackerName, $"ParseLatest completed successfully (took {sw.Elapsed.TotalSeconds:F1}s)");
+                }
+                catch (Exception ex)
+                {
+                    ParserLog.Write(TrackerName, $"ParseLatest Error: {ex.Message}");
                 }
 
-                ParserLog.Write(TrackerName, $"ParseLatest completed successfully (took {sw.Elapsed.TotalSeconds:F1}s)");
-            }
-            catch (Exception ex)
-            {
-                ParserLog.Write(TrackerName, $"ParseLatest Error: {ex.Message}");
-            }
-            finally
-            {
-                _parseLatestSemaphore.Release();
-            }
-
-            var logText = log.ToString();
-            return string.IsNullOrWhiteSpace(logText) ? "ok" : logText;
+                return log.ToString();
+            });
         }
 
         async Task<bool> parsePage(string cat, int page, string arg = null)
