+++ Infrastructure/Trackers/TrackerSyncHelpers.cs
@@ -1,10 +1,12 @@
 // Tracker sync shared helpers — parse lock and cron guard patterns.
 //
-// All *SyncService classes use TrackerParseLock + RunParseAsync for ParseAsync (and shared locks for API variants).
-// ParseAllTask / ParseLatest still use TrackerWorkFlag or SemaphoreSlim per tracker where applicable.
+// ParseAsync: TrackerParseLock + RunParseAsync
+// ParseAllTask: TrackerWorkFlag + RunParseAllTaskAsync
+// ParseLatest: TrackerLatestParseLock + RunParseLatestAsync
 
 using JacRed.Infrastructure.Logging;
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace JacRed.Infrastructure.Trackers
@@ -36,7 +38,7 @@ namespace JacRed.Infrastructure.Trackers
         }
     }
 
-    /// <summary>Simple volatile work flag for secondary jobs (e.g. ParseAllTask).</summary>
+    /// <summary>Work flag for secondary jobs (ParseAllTask).</summary>
     public sealed class TrackerWorkFlag
     {
         volatile bool _work;
@@ -44,6 +46,17 @@ namespace JacRed.Infrastructure.Trackers
         public bool TryStart() => System.Threading.Interlocked.CompareExchange(ref _work, true, false) == false;
 
         public void End() => System.Threading.Interlocked.Exchange(ref _work, false);
+    }
+
+    /// <summary>Semaphore guard for ParseLatest (one concurrent run per tracker).</summary>
+    public sealed class TrackerLatestParseLock
+    {
+        readonly SemaphoreSlim _semaphore = new SemaphoreSlim(1, 1);
+
+        public Task<bool> TryEnterAsync(CancellationToken cancellationToken = default)
+            => _semaphore.WaitAsync(0, cancellationToken);
+
+        public void Exit() => _semaphore.Release();
     }
 
     public static class TrackerSyncHelpers
@@ -62,10 +75,6 @@ namespace JacRed.Infrastructure.Trackers
             JacRedLog.Debug(JacRedLogCategories.Trackers, $"{trackerName}: parse skipped ({reason})");
         }
 
-        /// <summary>
-        /// Runs parse body when tracker is enabled and parse lock is free.
-        /// Returns <see cref="DisabledResult"/>, <see cref="WorkResult"/>, or the action result.
-        /// </summary>
         public static async Task<string> RunParseAsync(
             string trackerName,
             TrackerParseLock parseLock,
@@ -91,6 +100,73 @@ namespace JacRed.Infrastructure.Trackers
             finally
             {
                 parseLock.End();
+            }
+        }
+
+        public static async Task<string> RunParseAllTaskAsync(
+            string trackerName,
+            TrackerWorkFlag workFlag,
+            bool checkDisabled,
+            Func<Task> action,
+            CancellationToken cancellationToken = default)
+        {
+            if (checkDisabled && IsTrackerDisabled(trackerName))
+            {
+                LogParseSkipped(trackerName, DisabledResult);
+                return DisabledResult;
+            }
+
+            if (!workFlag.TryStart())
+            {
+                LogParseSkipped(trackerName, WorkResult);
+                return WorkResult;
+            }
+
+            try
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+                await action();
+            }
+            catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
+            {
+                throw;
+            }
+            catch { }
+            finally
+            {
+                workFlag.End();
+            }
+
+            return "ok";
+        }
+
+        public static async Task<string> RunParseLatestAsync(
+            string trackerName,
+            TrackerLatestParseLock latestLock,
+            bool checkDisabled,
+            Func<Task<string>> buildLogAsync,
+            CancellationToken cancellationToken = default)
+        {
+            if (checkDisabled && IsTrackerDisabled(trackerName))
+            {
+                LogParseSkipped(trackerName, DisabledResult);
+                return DisabledResult;
+            }
+
+            if (!await latestLock.TryEnterAsync(cancellationToken))
+            {
+                LogParseSkipped(trackerName, WorkResult);
+                return WorkResult;
+            }
+
+            try
+            {
+                var logText = await buildLogAsync();
+                return string.IsNullOrWhiteSpace(logText) ? "ok" : logText;
+            }
+            finally
+            {
+                latestLock.Exit();
             }
         }
     }
