--- Models/TorrentInfo.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace JacRed.Models
-{
-    public class TorrentInfo
-    {
-        public DateTime updateTime { get; set; }
-
-        public long fileTime { get; set; }
-    }
-}
